Moved some config tests out of test_io
